Move CrossChainWithdraw trait to traits module
