+++ primitives/polkadex/src/lib.rs
@@ -27,6 +27,7 @@ pub mod auction;
 pub mod fees;
 pub mod rewards;
 pub mod withdrawal;
+pub mod traits;
 
 pub use frame_support::storage::bounded_vec::BoundedVec;
